termionix_service

Enum TelnetOption

#[repr(u8)]
pub enum TelnetOption {

    TransmitBinary,
    Echo,
    Reconnection,
    SuppressGoAhead,
    NegotiateApproxMessageSize,
    Status,
    TimingMark,
    RCTE,
    OutLineWidth,
    OutPageSize,
    NAOCRD,
    NAOHTS,
    NAOHTD,
    NAOFFD,
    NAOVTS,
    NAOVTD,
    NAOLFD,
    XASCII,
    Logout,
    ByteMacro,
    DET,
    SUPDUP,
    SUPDUPOutput,
    SNDLOC,
    TTYPE,
    EOR,
    TUID,
    OUTMRK,
    TTYLOC,
    OPT3270Regime,
    X3PAD,
    NAWS,
    TSPEED,
    LFLOW,
    Linemode,
    XDISPLOC,
    Environment,
    Authentication,
    Encryption,
    NewEnvironment,
    TN3270E,
    XAUTH,
    Charset,
    TRSP,
    CPCO,
    TSLE,
    StartTLS,
    Kermit,
    SendUrl,
    ForwardX,
    MSDP,
    MSSP,
    Compress1,
    Compress2,
    ZMP,
    PragmaLogon,
    SSPILogon,
    PragmaHeartbeat,
    GMCP,
    EXOPL,
    Unknown(u8),

}

Telnet Terminal Options

Variants

TransmitBinary

[consts::option::BINARY] Telnet Binary Transmission RFC856

Echo

[consts::option::ECHO] Telnet Echo Option RFC857

Reconnection

[consts::option::RCP] Telnet Reconnection Option DDN Protocol Handbook, “Telnet Reconnection Option”, NIC 50005, December 1985. Note: Prepare to reconnect

SuppressGoAhead

[consts::option::SGA] Suppress Go ahead RFC858

NegotiateApproxMessageSize

[consts::option::NAMS] Negotiate Approximate Message Size

Status

[consts::option::STATUS] Telnet Status Option RFC859

TimingMark

[consts::option::TM] Telnet Timing Mark Option RFC860

RCTE

[consts::option::RCTE] Remote-Controlled Transmission and Echo RFC726

OutLineWidth

[consts::option::NAOL] Output Line Width DDN Protocol Handbook, “Telnet Output Line Width Option”, NIC 50005, December 1985.

OutPageSize

[consts::option::NAOP] Output Page Size DDN Protocol Handbook, “Telnet Output Page Size Option”, NIC 50005, December 1985.

NAOCRD

[consts::option::NAOCRD] Output Carriage-Return Disposition RFC652

NAOHTS

[consts::option::NAOHTS] Output Horizontal Tab Stops RFC653

NAOHTD

[consts::option::NAOHTD] Output Horizontal Tab Disposition RFC654

NAOFFD

[consts::option::NAOFFD] Output Form Feed Disposition RFC655

NAOVTS

[consts::option::NAOVTS] Output Vertical Tab Stops RFC656

NAOVTD

[consts::option::NAOVTD] Output Vertical Tab Disposition RFC657

NAOLFD

[consts::option::NAOLFD] Output Linefeed Disposition RFC658

XASCII

[consts::option::XASCII] Extended ASCII RFC698

Logout

[consts::option::LOGOUT] Logout Option RFC727

ByteMacro

[consts::option::BM] Byte Macro RFC735

DET

[consts::option::DET] Data Entry Terminal RFC1043 RFC732

SUPDUP

[consts::option::SUPDUP] SUPDUP RFC736 RFC734

SUPDUPOutput

[consts::option::SUPDUP_OUTPUT] SUPDUP Output RFC749

SNDLOC

[consts::option::SNDLOC] Send Location RFC779

TTYPE

[consts::option::TTYPE] Terminal Type RFC1091

EOR

[consts::option::EOR] End of Record RFC885

TUID

[consts::option::TUID] TACACS User Identification RFC927

OUTMRK

[consts::option::OUTMRK] Output Marking RFC933

TTYLOC

[consts::option::TTYLOC] Terminal Location Number RFC946

OPT3270Regime

[consts::option::OPT3270REGIME] Telnet 3270 Regime RFC1041

X3PAD

[consts::option::X3PAD] X.3 PAD RFC1053

NAWS

[consts::option::NAWS] Negotiate About Window Size RFC1073

TSPEED

[consts::option::TSPEED] Terminal Speed RFC1079

LFLOW

[consts::option::LFLOW] Remote Flow Control RFC1372

Linemode

[consts::option::LINEMODE] Linemode RFC1184

XDISPLOC

[consts::option::XDISPLOC] X Display Location RFC1096

Environment

[consts::option::OLD_ENVIRONMENT] Environment Option RFC1408

Authentication

[consts::option::AUTHENTICATION] Authentication Option RFC2941

Encryption

[consts::option::ENCRYPTION] Encryption Option RFC2946

NewEnvironment

[consts::option::NEW_ENVIRONMENT] New Environment Option RFC1572

TN3270E

[consts::option::TN3270E] TN3270E RFC2355

XAUTH

[consts::option::XAUTH] XAUTH Rob_Earhart

Charset

[consts::option::CHARSET] Charset RFC2066

TRSP

[consts::option::TRSP] Telnet Remote Serial Port (RSP) Robert_Barnes

CPCO

[consts::option::CPCO] Com Port Control Option RFC2217

TSLE

[consts::option::TSLE] Telnet Suppress Local Echo Wirt_Atmar

StartTLS

[consts::option::START_TLS] Telnet Start TLS Michael_Boe

Kermit

[consts::option::KERMIT] Kermit RFC2840

SendUrl

[consts::option::SENDURL] SEND-URL David_Croft

ForwardX

[consts::option::FORWARDX] FORWARD_X Jeffrey_Altman

MSDP

[consts::option::MSDP] Mud Server Data Protocol MSDP

MSSP

[consts::option::MSSP] Mud Server Status Protocol MSSP

Compress1

[consts::option::COMPRESS1] Mud Client Compression Protocol version 1 MCCPv1

Compress2

[consts::option::COMPRESS2] Mud Client Compression Protocol version 2 MCCPv2

ZMP

[consts::option::ZMP] Zenith Mud Protocol ZMP

PragmaLogon

[consts::option::PRAGMA_LOGIN] Telnet Option Pragma Logon Steve_McGregory

SSPILogon

[consts::option::SSPI_LOGIN] Telnet Option SSPI Logon Steve_McGregory

PragmaHeartbeat

[consts::option::PRAGMA_HEARTBEAT] Telnet Option Pragma Heartbeat Steve_McGregory

GMCP

[consts::option::GMCP] Generic Mud Communication Protocol GMCP Protocol

EXOPL

[consts::option::EXOPL] Extended-Options-List RFC861

Unknown(u8)

Unknown Option

Implementations

impl TelnetOption

pub fn to_u8(&self) -> u8

Converts a TelnetOption instance into its corresponding u8 representation.

Returns

This function maps each possible variant of the TelnetOption enum to its associated constant value defined under consts::option, or the raw byte value for the TelnetOption::Unknown variant. Each value corresponds to a specific Telnet sidechannel option code.

pub fn from_u8(byte: u8) -> TelnetOption

Converts a u8 value representing a Telnet option into the corresponding variant of the TelnetOption enum.

Arguments

byte - A u8 value that corresponds to a specific Telnet option as defined by the constants in consts::option.

Returns

Returns a variant of the TelnetOption enum corresponding to the provided byte. If an unknown or unsupported value is provided, the TelnetOption::Unknown(byte) variant is returned containing the original byte.

pub fn supported_local(&self) -> bool

Whether we support this option from us -> them.

pub fn supported_remote(&self) -> bool

Whether we support this option from them -> us.